+++ tests/e2e/tests/admin/logout.spec.js
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 // eslint-disable-next-line import/extensions
-import { resetDatabaseAndImportDataFromPath } from '../../scripts/dts-import';
+import { resetDatabaseAndImportDataFromPath } from '../../utils/dts-import';
 import { login } from '../../utils/login';
 
 test.describe('Log Out', () => {
