+++ tests/e2e/tests/admin/transfer/tokens.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { login } from '../../../utils/login';
-import { resetDatabaseAndImportDataFromPath } from '../../../scripts/dts-import';
+import { resetDatabaseAndImportDataFromPath } from '../../../utils/dts-import';
 import { navToHeader } from '../../../utils/shared';
 
 const createTransferToken = async (page, tokenName, duration, type) => {
