+++ tests/e2e/tests/content-manager/listview.spec.ts
@@ -3,7 +3,7 @@
  */
 import { test, expect } from '@playwright/test';
 import { login } from '../../utils/login';
-import { resetDatabaseAndImportDataFromPath } from '../../scripts/dts-import';
+import { resetDatabaseAndImportDataFromPath } from '../../utils/dts-import';
 
 test.describe('List View', () => {
   test.beforeEach(async ({ page }) => {
