+++ tests/e2e/tests/content-type-builder/tutorial.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 import { login } from '../../utils/login';
-import { resetDatabaseAndImportDataFromPath } from '../../scripts/dts-import';
+import { resetDatabaseAndImportDataFromPath } from '../../utils/dts-import';
 
 test.describe('Tutorial', () => {
   test.beforeEach(async ({ page }) => {
